+++ Examples/YouTubeClone/YTLiveChatViewController.swift
@@ -169,8 +169,8 @@ final class YTLiveChatViewController: _ViewController,
             .store(in: &cancellables)
 
         livestreamChat.state.$typingUsers
-            .sink { [weak self] typingUsers in
-                self?.applyTypingUsersUpdate(typingUsers)
+            .sink { [weak self] _ in
+                self?.applyTypingUsersUpdate()
             }
             .store(in: &cancellables)
     }
@@ -193,14 +193,13 @@ final class YTLiveChatViewController: _ViewController,
         messageListVC.updateMessages(with: changes)
     }
 
-    private func applyTypingUsersUpdate(_ typingUsers: Set<ChatUser>) {
+    private func applyTypingUsersUpdate() {
         guard livestreamChat.state.channel?.canSendTypingEvents == true else { return }
 
         let currentUserId = client.currentUserId
-        let typingUsersWithoutCurrentUser = typingUsers
-            .sorted { $0.id < $1.id }
+        let typingUsersWithoutCurrentUser = livestreamChat.state.typingUsersWithMemberInfo
             .filter { $0.id != currentUserId }
-            .map { TypingUser(user: $0, memberInfo: livestreamChat.state.typingMemberInfos[$0.id]) }
+            .sorted { $0.id < $1.id }
 
         if typingUsersWithoutCurrentUser.isEmpty {
             messageListVC.hideTypingIndicator()