+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Add `ChatMessage.mentionedChannelMembers` for accessing channel-member info of mentioned users [#4241](https://github.com/GetStream/stream-chat-swift/pull/4241)
 - Add `TypingUser` and `ChatChannel.typingUsers` for accessing channel-member info of typing users [#4241](https://github.com/GetStream/stream-chat-swift/pull/4241)
 - Add `LivestreamChatState.typingMemberInfos` for accessing channel-member info of typing users [#4241](https://github.com/GetStream/stream-chat-swift/pull/4241)
+- Add `LivestreamChatState.typingUsersWithMemberInfo` for `TypingUser` values derived from `typingUsers` and `typingMemberInfos` [#4241](https://github.com/GetStream/stream-chat-swift/pull/4241)
 ### 🐞 Fixed
 - Fix rare crashes in `DataController.state` [#4248](https://github.com/GetStream/stream-chat-swift/pull/4248)
 ### 🔄 Changed